# Refactor/job parsing performance