--- a/R/RSimpleQueue.R
+++ b/R/RSimpleQueue.R
@@ -143,7 +143,7 @@ RSimpleQueue <-
                   }
                   return(job_resp)
                 },
-                cancel_job = function(id) {
+                cancel_job = function(id, parse = TRUE) {
                   job_resp <- safe_put(glue("{address}/job/cancel/{id}", address = private$address))$result
                   if (parse) {
                     return(parse_response(job_resp))