--- a/inst/qview/.gitignore
+++ b/inst/qview/.gitignore
@@ -1,0 +1,2 @@
+skeleton.R
+proto.R