--- a/inst/qview/long_sample.R
+++ b/inst/qview/long_sample.R
@@ -1,0 +1,3 @@
+Sys.sleep(15)
+
+"<!> Look at this nifty return string! </!>"