This pull request refactors how jobs are parsed given noticeable performance implications when multiple observers are on a single R process


# performance characteristics

```r
TRAC: observation triggered for monitoring    
TRAC: getting jobs                            
TRAC: ongoing jobs                             ongoing=20:22
TRAC: newly completed jobs                     completed=19
TRAC: completed checking job status            duration=0.196717023849487
TRAC: observation triggered for monitoring    
TRAC: getting jobs                            
TRAC: ongoing jobs                             ongoing=20:22
TRAC: newly completed jobs                     completed=integer(0)
TRAC: completed checking job status            duration=0.126863956451416
```

Of key note, is the duration (in seconds) to run through the (naive) observe loop 
where all jobs are queried

Examination of the performance shows that the majority of time is spent aggregating
the jobs to the dataframe and converting to times in each iteration

![naive_flame_graph](https://user-images.githubusercontent.com/3196313/47440851-059f2600-d77d-11e8-9d06-122881f2bcb6.png)


rewriting the jobs_to_df function to optimize for the individual jobs
shows that an ~6x performance increase can be achieve by hand-coding the access
to fields of interest, though date-time conversion has not yet occurred. That said,
date-time conversion can now be done via a single vectorized invocation.

```
# A tibble: 3 x 14
  expression       min     mean   median      max `itr/sec` mem_alloc  n_gc n_itr 
  <chr>       <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl> <bch:byt> <dbl> <int> 
1 jdf(jobs[[…  779.4µs    1.5ms   1.36ms  35.31ms      664.      280B     4   307 
2 jdf2(jobs[… 516.39µs 745.49µs 715.51µs   4.77ms     1341.        0B     5   629 
3 jobs_to_df…   3.15ms   4.21ms   3.94ms   5.84ms      238.    30.4KB     6   104 
```

Second, the strategy for aggregating the individual jobs was evaluated to see
if `map_dfr` had any performance implications:

```r
bench::mark(
   map_dfr(jobs, jdf2),
   map(jobs, jdf2) %>% bind_rows(),
   lapply(jobs, jdf2) %>% bind_rows(),
   lapply(jobs, jdf2) %>% data.table::rbindlist(),
  check = FALSE
)

  expression    min   mean median    max `itr/sec` mem_alloc  n_gc n_itr
  <chr>      <bch:> <bch:> <bch:> <bch:>     <dbl> <bch:byt> <dbl> <int>
1 map_dfr(j… 19.6ms 20.9ms 20.6ms 23.3ms      47.8    7.31KB     6    16
2 map(jobs,… 19.5ms 21.9ms 22.4ms 23.6ms      45.7    6.71KB     5    16
3 lapply(jo… 19.1ms 28.2ms 22.6ms 77.2ms      35.5    6.71KB     5    13
4 lapply(jo… 16.6ms 18.3ms 18.5ms 21.2ms      54.6   30.52KB     6    19
```

This showed that performance is comparable, hence map_dfr will continue to be used



```{r eval = FALSE}
jdf <- function(.j) {
  data.frame(
    ID = .j$id,
    status = .j$status,
    context = .j$context,
    user = .j$user
  ) %>%
    cbind(as.data.frame(.j$run_details)) %>%
    cbind(.j$result)
}
jdf2 <- function(.j) {
  data.frame(
    ID = .j$id,
    status = .j$status,
    context = .j$context,
    user = .j$user,
    queue_time = .j$run_details$queue_time,
    start_time = .j$run_details$start_time,
    end_time = .j$run_details$end_time,
    exit_code = .j$result$exit_code,
    output = .j$result$output
  )
}
```

```{r eval = FALSE}
 lgr$trace("observation triggered for monitoring")
    if (!is.null(jobs$monitoring)) {
    start_time <- Sys.time()
    lgr$trace("getting jobs")
      job_df <- rsq$get_jobs() %>% jobs_to_df()
      ongoing <- job_df %>% filter(!(STATUS %in% c("COMPLETED", "ERROR"))) %>% pull(ID)
    lgr$with_fields(ongoing = ongoing)$trace("ongoing jobs")
      newly_completed <- setdiff(jobs$monitoring, ongoing)
    lgr$with_fields(completed = newly_completed)$trace("newly completed jobs")
      if (length(newly_completed)) {
        lapply(newly_completed, function(ncj) {
          cj <- job_df %>% filter(ID == ncj)
          if (cj$STATUS == "COMPLETED") {
            toastr_success(glue::glue("analysis completed: {cj$CONTEXT}"), position = "bottom-right" )
          } else {
            toastr_error(glue::glue("analysis failed: {cj$CONTEXT}"), position = "bottom-right" )
          }
        })
      }
      if (length(ongoing)) {
        message("still monitoring jobs: ", glue::glue_collapse(ongoing, sep = ","))
        jobs$monitoring <- ongoing
        invalidateLater(1000, session)
      } else {
        jobs$monitoring <- NULL
        toastr_success(glue::glue("all requested analyses completed"), position = "bottom-right" )
        message("completed monitoring, we're done")
      }
    lgr$with_fields(duration = diff_time(start_time))$trace("completed checking job status")
    }
```