@dpastoor 2018-10-24
On refactoring, we can see a drop from ~ 150ms to 30 ms, making it much more suitable when polling frequently

![image](https://user-images.githubusercontent.com/3196313/47441151-8d853000-d77d-11e8-9646-9e003d1f734c.png)